requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

python-multipart==0.0.9
PyPDF2==3.0.1

google-generativeai==0.8.5
python-dotenv==1.0.1

pydantic==2.6.4
requests==2.31.0
httpx==0.26.0

# required
spacy>=3.5.0
dateparser>=1.1.8
rapidfuzz>=2.15.1



